LiteLLM-Proxy/requirements.txt
litellm[proxy]
boto3
ell-ai